# Conflicts:
#	features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/common/util/ResourceProvider.kt
#	features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/detail/presentation/fragment/AddEditProductDetailFragment.kt
#	features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/detail/presentation/viewmodel/AddEditProductDetailViewModel.kt
#	features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/preview/presentation/service/AddEditProductEditService.kt
#	features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/preview/presentation/viewmodel/AddEditProductPreviewViewModel.kt
#	features/merchant/product_add_edit/src/test/java/com/tokopedia/product/addedit/detail/presentation/viewmodel/AddEditProductDetailViewModelTest.kt
#	features/merchant/product_add_edit/src/test/java/com/tokopedia/product/addedit/preview/presentation/viewmodel/AddEditProductPreviewViewModelTest.kt
#	features/merchant/product_add_edit/src/test/java/com/tokopedia/product/addedit/preview/presentation/viewmodel/AddEditProductPreviewViewModelTestFixture.kt
#	features/merchant/product_manage_common/src/main/res/layout/fragment_quick_edit_stock.xml
#	features/merchant/product_manage_list/src/main/res/layout/item_campaign_stock_total_editor.xml
#	features/merchant/shop_common/src/main/java/com/tokopedia/shop/common/constant/AccessId.kt
#	features/merchant/shop_common/src/main/java/com/tokopedia/shop/common/data/source/cloud/query/AuthorizeAccess.kt
#	features/merchant/shop_common/src/main/java/com/tokopedia/shop/common/domain/interactor/AuthorizeAccessUseCase.kt
#	features/transaction/sellerorder/src/main/java/com/tokopedia/sellerorder/list/presentation/fragments/SomListFragment.kt

+++ features/merchant/review/src/main/java/com/tokopedia/review/common/domain/usecase/ProductrevGetReviewDetailUseCase.kt
@@ -55,12 +55,12 @@ class ProductrevGetReviewDetailUseCase @Inject constructor(graphqlRepository: Gr
     }
 
     @GqlQuery(REVIEW_DETAIL_QUERY_CLASS_NAME, REVIEW_DETAIL_QUERY)
-    fun setRequestParams(feedbackID: Int) {
+    fun setRequestParams(feedbackID: Long) {
         setGraphqlQuery(ReviewDetail.GQL_QUERY)
         setTypeClass(ProductrevGetReviewDetailResponseWrapper::class.java)
         setRequestParams(
                 RequestParams.create().apply {
-                    putInt(PARAM_FEEDBACK_ID, feedbackID)
+                    putLong(PARAM_FEEDBACK_ID, feedbackID)
                 }.parameters
         )
     }
